Improve initial import
