+++ earthkit/data/sources/ecmwf_api.py
@@ -11,7 +11,6 @@ import logging
 
 from earthkit.data.core.thread import SoftThreadPool
 from earthkit.data.decorators import normalize
-from earthkit.data.utils import tqdm
 
 from .file import FileSource
 from .prompt import APIKeyPrompt
@@ -65,6 +64,8 @@ class ECMWFApi(FileSource):
         if nthreads < 2:
             self.path = [self._retrieve(r) for r in requests]
         else:
+            from earthkit.data.utils.progbar import tqdm
+
             with SoftThreadPool(nthreads=nthreads) as pool:
                 futures = [pool.submit(self._retrieve, r) for r in requests]
