+++ earthkit/data/sources/wekeo.py
@@ -15,7 +15,6 @@ import yaml
 from hda.api import DataOrderRequest
 
 from earthkit.data.core.thread import SoftThreadPool
-from earthkit.data.utils import tqdm
 
 from .file import FileSource
 from .prompt import APIKeyPrompt
@@ -122,6 +121,8 @@ class WekeoRetriever(FileSource):
         if nthreads < 2:
             self.path = [self._retrieve(dataset, r) for r in requests]
         else:
+            from earthkit.data.utils.progbar import tqdm
+
             with SoftThreadPool(nthreads=nthreads) as pool:
                 futures = [pool.submit(self._retrieve, dataset, r) for r in requests]
