+++ environment.yml
@@ -29,7 +29,7 @@ dependencies:
   - earthkit-meteo>=0.0.1
   - eccovjson>=0.0.5
   - earthkit-geo
-- tqdm
+- tqdm>=4.63.0
 - markdown
 - make
 - mypy
