+++ setup.cfg
@@ -32,7 +32,7 @@ install_requires =
     netcdf4
     pandas
     pyyaml
-    tqdm
+    tqdm>=4.63.0
     xarray>=0.19.0
     earthkit-meteo>=0.0.1
     earthkit-geo
