+++ tests/downstream-ci-requirements.txt
@@ -18,7 +18,7 @@ pyfdb
 pyodc
 pyyaml
 torch
-tqdm
+tqdm>=4.63.0
 xarray>=0.19.0
 earthkit-meteo>=0.0.1
 earthkit-geo
